+++ Intent.Modules.Java.Services/Templates/Enum/EnumTemplatePartial.cs
@@ -12,13 +12,13 @@ using Intent.Templates;
 [assembly: DefaultIntentManaged(Mode.Merge)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.Java.Templates.JavaFileTemplatePartial", Version = "1.0")]
 
-namespace Intent.Modules.Java.Domain.Templates.Enum
+namespace Intent.Modules.Java.Services.Templates.Enum
 {
     [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
     partial class EnumTemplate : JavaTemplateBase<Intent.Modules.Common.Types.Api.EnumModel>
     {
         [IntentManaged(Mode.Fully)]
-        public const string TemplateId = "Intent.Java.Domain.Enum";
+        public const string TemplateId = "Intent.Java.Services.Enum";
 
         [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
         public EnumTemplate(IOutputTarget outputTarget, Intent.Modules.Common.Types.Api.EnumModel model) : base(TemplateId, outputTarget, model)
@@ -33,21 +33,6 @@ namespace Intent.Modules.Java.Domain.Templates.Enum
                 package: this.GetPackage(),
                 relativeLocation: this.GetFolderPath()
             );
-        }
-
-        private static string GetEnumLiterals(IEnumerable<EnumLiteralModel> literals)
-        {
-            return string.Join(@",
-    ", literals.Select(GetEnumLiteral));
-        }
-
-        private static string GetEnumLiteral(EnumLiteralModel literal)
-        {
-            // Java Enums don't have the literal values like C#, etc.
-            // If you want to store the literal values, you need a constructor
-            // and a field that can store int's.
-            // Example: https://www.baeldung.com/a-guide-to-java-enums#fields-methods-and-constructors-in-enums
-            return $"{literal.Name.ToJavaIdentifier(CapitalizationBehaviour.AsIs)}";
         }
     }
 }