+++ Intent.Modules.Java.Services/Templates/TemplateExtensions.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Java.Services.Templates.DataTransferModel;
+using Intent.Modules.Java.Services.Templates.Enum;
 using Intent.Modules.Java.Services.Templates.ExceptionType;
 using Intent.Modules.Java.Services.Templates.ServiceImplementation;
 using Intent.Modules.Java.Services.Templates.ServiceInterface;
@@ -21,6 +22,16 @@ namespace Intent.Modules.Java.Services.Templates
         public static string GetDataTransferModelName(this IntentTemplateBase template, Intent.Modelers.Services.Api.DTOModel model)
         {
             return template.GetTypeName(DataTransferModelTemplate.TemplateId, model);
+        }
+
+        public static string GetEnumName<T>(this IntentTemplateBase<T> template) where T : Intent.Modules.Common.Types.Api.EnumModel
+        {
+            return template.GetTypeName(EnumTemplate.TemplateId, template.Model);
+        }
+
+        public static string GetEnumName(this IntentTemplateBase template, Intent.Modules.Common.Types.Api.EnumModel model)
+        {
+            return template.GetTypeName(EnumTemplate.TemplateId, model);
         }
 
         public static string GetExceptionTypeName<T>(this IntentTemplateBase<T> template) where T : Intent.Modules.Java.Services.Api.ExceptionTypeModel